# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Video Streaming Service API
  version: 1.0.0
  contact:
    name: video-streaming-service
    url: https://github.com/Nuu-maan/video-streaming-service
  description: |
    Standalone JSON API for a video platform: accounts, uploads, transcoding,
    HLS/MP4 playback, comments, playlists, subscriptions, search, moderation
    and admin analytics. It is designed to be hosted on its own and consumed
    by a separate frontend (React, Next.js, Vue, ...) running on a different
    origin.

    ## The three things to know first

    1. **Auth is JWT Bearer.** Log in at `POST /auth/login` (the body field is
       `identifier` — a username *or* an email — plus `password`) and send
       `Authorization: Bearer <access_token>` on every authenticated call.
       Access tokens live 900 s; exchange the `refresh_token` at
       `POST /auth/refresh` for a new pair. A refresh token is *not* usable as
       an API credential, and an access token is rejected at `/auth/refresh`.

    2. **Every JSON response is wrapped in an envelope.**
       Success: `{"success": true, "data": ...}`.
       Lists (including `/search`): `{"success": true, "data": [...], "pagination": {"total", "page", "limit", "total_pages", "has_next", "has_previous"}}`.
       Errors: `{"success": false, "error": {"code": "...", "message": "..."}}`.
       Media routes (`.m3u8`, `.ts`, MP4, JPEG) return raw bytes, never the envelope.

...

    ## Other conventions

...
servers:
...
    variables:
...
security: []
paths:
...
  /videos/{id}/view:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Engagement]
      operationId: recordView
      summary: Record one view (explicit — playback does not auto-count)
      description: >-
        Views are deduplicated in Redis. Auth is optional: authenticated
        callers are deduped by user id; **anonymous callers must send a
        `session_id`**. The body must be JSON, even if just `{}` for an
        authenticated caller. **201 = counted, 200 = deduplicated repeat.**
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quality:
                  type: string
                source:
                  type: string
                session_id:
                  type: string
                  description: Required for anonymous callers
      responses:
        "201":
          description: View counted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ViewResult"
        "200":
          description: Duplicate within the dedupe window — not counted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ViewResult"
        "400":
          description: Missing `session_id` for an anonymous caller, or non-JSON body (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /videos/{id}/progress:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Engagement]
      operationId: saveProgress
      summary: Upsert the caller's resume position
      description: >-
        `position` and `duration` are non-negative seconds and **`position`
        must lie within `duration`**, otherwise the request is a 400.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: integer
                  minimum: 0
                  description: Seconds; must not exceed `duration`
                duration:
                  type: integer
                  minimum: 0
                  description: Seconds
                completed:
                  type: boolean
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Negative values, or `position` beyond `duration` (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /me/history:
    get:
      tags: [Engagement]
      operationId: getHistory
      summary: Watch history, most recently watched first
      description: An empty history serialises as `[]`, never `null`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Page of history entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/WatchHistory"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Engagement]
      operationId: clearHistory
      summary: Delete the caller's entire watch history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: History cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/history/{videoId}:
    delete:
      tags: [Engagement]
      operationId: deleteHistoryEntry
      summary: Remove one video from the caller's watch history
      security:
        - bearerAuth: []
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entry removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────── Streaming ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <access_token>`. Obtain a token pair from
        `/auth/login` or `/auth/register`; renew with `/auth/refresh`.
        Access tokens live 900 s. A refresh token is not accepted here.
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number; values below 1 are clamped to 1
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Page size; clamped to at most 100
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    VideoId:
      name: id
      in: path
      required: true
      description: Video id
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Malformed input (`VALIDATION_ERROR`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request
    Unauthorized:
      description: Missing, invalid, expired or revoked access token (`UNAUTHORIZED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
...
    NotFound:
      description: >-
        Resource not found (`NOT_FOUND`). Also the answer for a private video
        or playlist read by anyone but its owner — never 403.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Not found
  schemas:
...
    SuccessEnvelope:
      type: object
      description: 'Every non-media, non-list success response: `{"success": true, "data": ...}`'
      required: [success]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: The payload; its shape is documented per endpoint
    PaginatedEnvelope:
      type: object
      description: >-
        Every list endpoint — including /search — wraps its page in this
        envelope.
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items: {}
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
    ErrorResponse:
      type: object
      description: 'Every error: `{"success": false, "error": {"code", "message"}}`'
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            Machine-readable error code. `AUTH_UNAVAILABLE` arrives with
            status 503 when the token-revocation store is unreachable and the
            server fails closed — retry with the same token rather than
            discarding credentials as you would on a 401.
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - INTERNAL_ERROR
            - ALREADY_EXISTS
            - USER_BANNED
            - DUPLICATE_REPORT
            - EMAIL_ALREADY_VERIFIED
            - INVALID_TOKEN
            - INVALID_CURRENT_PASSWORD
            - FILE_TOO_LARGE
            - INVALID_FORMAT
            - HLS_NOT_READY
            - PLAYLIST_NOT_FOUND
            - SEGMENT_NOT_FOUND
            - VIDEO_NOT_READY
            - FILE_NOT_FOUND
            - ALREADY_IN_PLAYLIST
            - AUTH_UNAVAILABLE
        message:
          type: string
          description: Human-readable explanation
    MessageResponse:
      description: A success envelope whose payload is a short confirmation
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
              additionalProperties: true

    # ── Enums ──
...
    ViewResult:
      type: object
      properties:
        counted:
          type: boolean
          description: true when the view was counted (201); false on a deduplicated repeat (200)

    # ── Social ──
...
    WatchHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        video_id:
          type: string
          format: uuid
        watched_at:
          type: string
          format: date-time
        watch_duration:
          type: integer
        completed:
          type: boolean
        last_position:
          type: integer
...