# internal/handler/view_handler.go
package handler

import (
	"errors"
	"net/http"
	"strings"

	"github.com/gin-gonic/gin"

	"github.com/Nuu-maan/video-streaming-service/internal/domain"
	"github.com/Nuu-maan/video-streaming-service/internal/service"
	"github.com/Nuu-maan/video-streaming-service/pkg/appctx"
	"github.com/Nuu-maan/video-streaming-service/pkg/logger"
	"github.com/Nuu-maan/video-streaming-service/pkg/response"
	"github.com/Nuu-maan/video-streaming-service/pkg/validator"
)

// ViewHandler exposes view counting, playback progress, and watch history.
//
// Views are counted by an explicit POST from the player rather than from
// inside the HLS handlers, for two reasons: hls.js refetches a media playlist
// every few seconds, so counting playlist hits would count one viewer many
// times over; and playlists are cacheable/CDN-able, so a cached hit would
// count many viewers zero times. An explicit endpoint is also what an external
// frontend on another origin needs anyway.
type ViewHandler struct {
	tracker *service.ViewTracker
	log     *logger.Logger
}

func NewViewHandler(tracker *service.ViewTracker, log *logger.Logger) *ViewHandler {
	return &ViewHandler{tracker: tracker, log: log}
}

type recordViewRequest struct {
	Quality string `json:"quality"`
	Source  string `json:"source"`
	// SessionID is generated by the client and identifies an anonymous viewer
	// across requests; without it anonymous views could not be deduplicated.
	SessionID string `json:"session_id"`
}

type saveProgressRequest struct {
	Position  int  `json:"position"`
	Duration  int  `json:"duration"`
	Completed bool `json:"completed"`
}

// RecordView counts one view of a video. Anonymous viewers count too — they
// are identified by their client-generated session_id — and repeat plays by
// the same viewer inside the dedupe window do not count twice.
func (h *ViewHandler) RecordView(c *gin.Context) {
	ctx := c.Request.Context()

	videoID, err := validator.ValidateUUID(c.Param("id"))
	if err != nil {
		response.ValidationError(c, "Invalid video ID")
		return
	}

	var req recordViewRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.ValidationError(c, "request body must be JSON")
		return
	}

	userAgent := c.GetHeader("User-Agent")
	event := service.ViewEvent{
		VideoID:   videoID,
		SessionID: strings.TrimSpace(req.SessionID),
		// ClientIP honours X-Forwarded-For only for peers listed in gin's
		// TrustedProxies. That matters because the header is entirely
		// attacker-controlled: without a trusted-proxy list configured this
		// falls back to the socket peer address and spoofed forwarding headers
		// are ignored.
		IPAddress:  c.ClientIP(),
		UserAgent:  userAgent,
		Quality:    strings.TrimSpace(req.Quality),
		DeviceType: deviceTypeFrom(userAgent),
		// Country stays empty: resolving it needs a GeoIP database the service
		// does not ship, and the column is nullable for exactly that case.
		Source: strings.TrimSpace(req.Source),
	}

	if principal, ok := appctx.PrincipalFrom(ctx); ok {
		userID := principal.UserID
		event.UserID = &userID
	} else if event.SessionID == "" {
		response.ValidationError(c, "session_id is required for anonymous views")
		return
	}

	counted, err := h.tracker.TrackView(ctx, event)
	if err != nil {
		switch {
		case errors.Is(err, domain.ErrVideoNotFound):
			response.NotFound(c, "Video not found")
		case errors.Is(err, domain.ErrInvalidInput):
			response.ValidationError(c, "session_id is required for anonymous views")
		default:
			h.log.Error(ctx, "failed to record view", err, map[string]interface{}{"video_id": videoID})
			response.InternalError(c, "Failed to record view")
		}
		return
	}

	status := http.StatusCreated
	if !counted {
		// A deduplicated repeat is not an error — the player did nothing wrong
		// — it just did not create a view.
		status = http.StatusOK
	}
	response.Success(c, status, gin.H{"counted": counted})
}

// SaveProgress upserts the caller's resume position for a video.
func (h *ViewHandler) SaveProgress(c *gin.Context) {
	ctx := c.Request.Context()

	videoID, err := validator.ValidateUUID(c.Param("id"))
	if err != nil {
		response.ValidationError(c, "Invalid video ID")
		return
	}

	principal, ok := appctx.PrincipalFrom(ctx)
	if !ok {
		response.Unauthorized(c, "Authentication required")
		return
	}

	var req saveProgressRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.ValidationError(c, "position and duration are required")
		return
	}

	if err := h.tracker.SaveProgress(ctx, principal.UserID, videoID, req.Position, req.Duration, req.Completed); err != nil {
		switch {
		case errors.Is(err, domain.ErrVideoNotFound):
			response.NotFound(c, "Video not found")
		case errors.Is(err, domain.ErrInvalidProgress), errors.Is(err, domain.ErrInvalidInput):
			response.ValidationError(c, "position and duration must be non-negative seconds, with position within duration")
		default:
			h.log.Error(ctx, "failed to save watch progress", err, map[string]interface{}{"video_id": videoID})
			response.InternalError(c, "Failed to save progress")
		}
		return
	}

	response.Success(c, http.StatusOK, gin.H{"message": "Progress saved"})
}

// GetHistory returns the caller's watch history, most recently watched first.
func (h *ViewHandler) GetHistory(c *gin.Context) {
	ctx := c.Request.Context()

	principal, ok := appctx.PrincipalFrom(ctx)
	if !ok {
		response.Unauthorized(c, "Authentication required")
		return
	}

	page := parsePage(c)

	entries, total, err := h.tracker.History(ctx, principal.UserID, page.Limit, page.Offset)
	if err != nil {
		h.log.Error(ctx, "failed to load watch history", err, nil)
		response.InternalError(c, "Failed to load watch history")
		return
	}

	// An empty history must serialise as [] rather than null, or frontends end
	// up special-casing the first-ever request.
	if entries == nil {
		entries = []*domain.WatchHistory{}
	}

	response.SuccessWithList(c, entries, paginationMeta(total, page))
}

// ClearHistory deletes the caller's entire watch history.
func (h *ViewHandler) ClearHistory(c *gin.Context) {
	ctx := c.Request.Context()

	principal, ok := appctx.PrincipalFrom(ctx)
	if !ok {
		response.Unauthorized(c, "Authentication required")
		return
	}

	if err := h.tracker.ClearHistory(ctx, principal.UserID); err != nil {
		h.log.Error(ctx, "failed to clear watch history", err, nil)
		response.InternalError(c, "Failed to clear watch history")
		return
	}

	response.Success(c, http.StatusOK, gin.H{"message": "Watch history cleared"})
}

// DeleteHistoryEntry removes one video from the caller's watch history.
func (h *ViewHandler) DeleteHistoryEntry(c *gin.Context) {
	ctx := c.Request.Context()

	videoID, err := validator.ValidateUUID(c.Param("videoId"))
	if err != nil {
		response.ValidationError(c, "Invalid video ID")
		return
	}

	principal, ok := appctx.PrincipalFrom(ctx)
	if !ok {
		response.Unauthorized(c, "Authentication required")
		return
	}

	if err := h.tracker.RemoveHistoryEntry(ctx, principal.UserID, videoID); err != nil {
		if errors.Is(err, domain.ErrWatchHistoryNotFound) {
			response.NotFound(c, "Watch history entry not found")
			return
		}
		h.log.Error(ctx, "failed to delete watch history entry", err, map[string]interface{}{"video_id": videoID})
		response.InternalError(c, "Failed to delete watch history entry")
		return
	}

	response.Success(c, http.StatusOK, gin.H{"message": "Watch history entry removed"})
}

// deviceTypeFrom buckets a User-Agent into the three classes the analytics
// schema groups by. It is deliberately coarse: anything finer needs a real UA
// parser, and the analytics only ever segment mobile/desktop/tablet.
func deviceTypeFrom(userAgent string) string {
	ua := strings.ToLower(userAgent)
	switch {
	case ua == "":
		return ""
	case strings.Contains(ua, "ipad"), strings.Contains(ua, "tablet"),
		strings.Contains(ua, "android") && !strings.Contains(ua, "mobi"):
		return "tablet"
	case strings.Contains(ua, "mobi"), strings.Contains(ua, "android"):
		return "mobile"
	default:
		return "desktop"
	}
}